# api/openapispec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/variable-labels": {
            "get": {
                "description": "List variable labels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable_labels"
                ],
                "summary": "List variable labels",
                "operationId": "listVariableLabels",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Variable labels to filter variables by. Default to all(empty) labels.",
                        "name": "labels",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number of the paginated list results. Default to 1.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page size of the paginated list results. If not set, the result won't be paginated.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.VariableLabelsListResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new set of variable labels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable_labels"
                ],
                "summary": "Create variable labels",
                "operationId": "createVariableLabels",
                "parameters": [
                    {
                        "description": "Created variable labels",
                        "name": "variable_labels",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateVariableLabelsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.VariableLabels"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/variable-labels/{key}": {
            "get": {
                "description": "Get variable labels by variable key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable_labels"
                ],
                "summary": "Get variable labels",
                "operationId": "getVariableLabels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variable Key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.VariableLabels"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update the specified variable labels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable_labels"
                ],
                "summary": "Update variable labels",
                "operationId": "updateVariableLabels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variable Key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Variable Labels",
                        "name": "variable_labels",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateVariableLabelsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.VariableLabels"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete a specified variable with labels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable_labels"
                ],
                "summary": "Delete variable labels",
                "operationId": "deleteVariableLabels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variable Key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.VariableLabels": {
            "type": "object",
            "properties": {
                "labels": {
                    "description": "Labels is the list of variable labels, which should be sorted\nin ascending order of priority.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "variableKey": {
                    "description": "VariableKey is the access path for the variable.",
                    "type": "string"
                }
            }
        },
        "entity.VariableLabelsListResult": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "variableLabels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.VariableLabels"
                    }
                }
            }
        },
...
        "request.CreateVariableLabelsRequest": {
            "type": "object",
            "required": [
                "labels",
                "variableKey"
            ],
            "properties": {
                "labels": {
                    "description": "Labels is the list of variable labels, which should be sorted\nin ascending order of priority.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "variableKey": {
                    "description": "VariableKey is the access path for the variable.",
                    "type": "string"
                }
            }
        },
...
        "request.UpdateVariableLabelsRequest": {
            "type": "object",
            "required": [
                "labels",
                "variableKey"
            ],
            "properties": {
                "labels": {
                    "description": "Labels is the list of variable labels, which should be sorted\nin ascending order of priority.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "variableKey": {
                    "description": "VariableKey is the access path for the variable.",
                    "type": "string"
                }
            }
        },
...
    }
}